move monitor_v1_service to core
