+++ cmd/keytransparency-monitor/main.go
@@ -42,8 +42,8 @@ import (
 
 	cmon "github.com/google/keytransparency/core/monitor"
 	pb "github.com/google/keytransparency/core/proto/keytransparency_v1"
+	mopb "github.com/google/keytransparency/core/proto/monitor_v1_service"
 	mupb "github.com/google/keytransparency/core/proto/mutation_v1_service"
-	mopb "github.com/google/keytransparency/impl/proto/monitor_v1_service"
 	tlogcli "github.com/google/trillian/client"
 	_ "github.com/google/trillian/merkle/coniks"    // Register coniks
 	_ "github.com/google/trillian/merkle/objhasher" // Register objhasher
